third_party/blink/web_tests/storage/indexeddb/events-expected.txt

Test IndexedDB's event interfaces.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 'IDBVersionChangeEvent' in self is true
PASS 'oldVersion' in new IDBVersionChangeEvent('versionchange') is true
PASS 'newVersion' in new IDBVersionChangeEvent('versionchange') is true
PASS 'dataLoss' in new IDBVersionChangeEvent('versionchange') is true
PASS successfullyParsed is true

TEST COMPLETE
